+++ src/runtime.ts
@@ -1,5 +1,6 @@
 /* tslint:disable */
 /* eslint-disable */
+import { isAffinityError } from "./errors";
 /**
  * Affinity API
  * Affinity API for software platforms connecting practices to the compounder network.
@@ -227,6 +228,7 @@ export class BaseAPI {
         }
       }
       if (response === undefined) {
+        if (isAffinityError(e)) throw e;
         if (e instanceof Error) {
           throw new FetchError(
             e,